allow generation type or module changes in eval pipeline
