+++ nemo_skills/inference/__init__.py
@@ -11,3 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .factory import GenerationType, GENERATION_MODULE_MAP
+
+__all__ = ["GenerationType", "GENERATION_MODULE_MAP"]
