+++ nemo_skills/pipeline/utils/eval.py
@@ -15,13 +15,13 @@
 import importlib
 import logging
 import os
-from collections import defaultdict
 from copy import deepcopy
 from dataclasses import dataclass, field
 from pathlib import Path
 
 import nemo_skills.pipeline.utils as pipeline_utils
 from nemo_skills.dataset.utils import get_dataset_module
+from nemo_skills.inference import GENERATION_MODULE_MAP
 from nemo_skills.inference.generate import GenerationTask
 from nemo_skills.utils import compute_chunk_ids, get_logger_name
 
@@ -218,10 +218,18 @@ def prepare_eval_commands(
     with_sandbox,
     wandb_parameters,
     extra_eval_args,
+    generation_type=None,
+    generation_module=None,
 ):
     # TODO: there is a bit too much code duplication here and logic is quite dense, should try to refactor
 
     # TODO: should we allow setting num chunks per benchmark when not using groups? Maybe benchmark:rs_num:num_chunks?
+
+    if generation_type is not None:
+        if generation_module is not None:
+            raise ValueError("Cannot specify both generation_module and generation_type. ")
+        
+        generation_module = GENERATION_MODULE_MAP[generation_type]
 
     benchmarks_or_groups = {
         k: int(v) for k, v in [b.split(":") if ":" in b else (b, -1) for b in benchmarks_or_groups.split(",")]
@@ -338,10 +346,10 @@ def prepare_eval_commands(
             for chunk_id in benchmark_chunk_ids:
                 job_benchmarks.add(benchmark)
 
-                generation_task = importlib.import_module(benchmark_args.generation_module)
+                generation_task = importlib.import_module(generation_module or benchmark_args.generation_module)
                 if not hasattr(generation_task, 'GENERATION_TASK_CLASS'):
                     raise ValueError(
-                        f"Module {benchmark_args.generation_module} does not have a GENERATION_TASK_CLASS attribute. "
+                        f"Module {generation_module or benchmark_args.generation_module} does not have a GENERATION_TASK_CLASS attribute. "
                         "Please provide a valid generation module."
                     )
                 generation_task = generation_task.GENERATION_TASK_CLASS
